src_student/app/requirements.txt
flask==2.1.0
uWSGI==2.0.19.1
sentry-sdk[flask]
python-logstash
